+++ src/MentionManager.hpp
@@ -5,7 +5,6 @@
 #include <Geode/utils/StringMap.hpp>
 #include <Geode/utils/async.hpp>
 
-#include <chrono>
 #include <string>
 #include <vector>
 
@@ -42,8 +41,5 @@ private:
 
     CommentObject formatCommentObj(const std::string& str);
 
-    bool m_doWeHaveInternet = true;
-    std::chrono::steady_clock::time_point m_nextInternetCheck =
-        std::chrono::steady_clock::now();
-    arc::Future<bool> doWeHaveInternet();
+    bool m_didWeHaveInternet = true;
 };